+++ praetorian_cli/sdk/entities/agents.py
@@ -8,28 +8,28 @@ class Agents:
     def __init__(self, api):
         self.api = api
 
-    def attribution(self, risk_key: str) -> str:
-        self.api.agent(AgentType.ATTRIBUTION.value, dict(key=risk_key))
+    def affiliation(self, key, timeout=180) -> str:
+        self.api.agent(AgentType.AFFILIATION.value, dict(key=key))
 
         # poll for the attribution job to complete
-        job_key = self.api.jobs.system_job_key(AgentType.ATTRIBUTION.value, risk_key)
+        job_key = self.api.jobs.system_job_key(AgentType.AFFILIATION.value, key)
 
         start_time = time()
-        while time() - start_time < 180:
+        while time() - start_time < timeout:
             job = self.api.jobs.get(job_key)
             if self.api.jobs.is_failed(job):
-                raise Exception('Attribution failed')
+                raise Exception('Failed to retrieve affiliation data.')
             if self.api.jobs.is_passed(job):
                 break
             sleep(1)
 
         if self.api.jobs.is_passed(job):
-            return self.attribution_result(risk_key)
+            return self.affiliation_result(key)
         else:
-            raise Exception('Timeout (3 minutes) waiting for attribution result.')
+            raise Exception(f'Timeout waiting for affiliation result ({timeout} seconds).')
 
-    def attribution_filename(self, agent_type: str, risk_key: str) -> str:
-        return f'agents/{agent_type}/{risk_key}'
+    def affiliation_filename(self, agent_type: str, key: str) -> str:
+        return f'agents/{agent_type}/{key}'
 
-    def attribution_result(self, risk_key: str) -> dict:
-        return self.api.files.get(self.attribution_filename(AgentType.ATTRIBUTION.value, risk_key)).decode('utf-8')
+    def affiliation_result(self, key: str) -> dict:
+        return self.api.files.get(self.affiliation_filename(AgentType.AFFILIATION.value, key)).decode('utf-8')