+++ praetorian_cli/sdk/test/test_z_cli.py
@@ -280,7 +280,7 @@ class TestZCli:
         self.verify('purge --help', ignore_stdout=True)
 
         self.verify('agent --help', ignore_stdout=True)
-        self.verify('agent attribution --help', ignore_stdout=True)
+        self.verify('agent affiliation --help', ignore_stdout=True)
 
     def verify(self, command, expected_stdout=[], expected_stderr=[], ignore_stdout=False):
         result = run(f'praetorian --profile "{self.sdk.keychain.profile}" chariot {command}', capture_output=True,